% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_.R
\name{anova_}
\alias{anova_}
\title{ANOVA tables with the original object as an attribute}
\usage{
anova_(object, ...)
}
\arguments{
\item{object}{An object for which anova or deviance tables should be computed.}

\item{...}{Additional arguments passed to the \code{anova()} function.}
}
\value{
An anova object with an additional \code{"object"} attribute containing
the original input.
}
\description{
This function attempts to compute anova or deviance tables using the
standard \code{\link[stats:anova]{stats::anova()}} function. The original object is attached as an
attribute to the result for reference.
}
\examples{
is_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
anova(is_lm)

anova_(is_lm)
attr(anova_(is_lm), "object")
}
