% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.function.lm.R
\name{as.function.lm}
\alias{as.function.lm}
\title{Transform an lm or glm model into a function}
\usage{
\method{as.function}{lm}(x, ...)
}
\arguments{
\item{x}{An lm or glm model}

\item{...}{Further arguments to the method (not used for now)}
}
\value{
A function with argument \code{x} that returns the values predicted by the
model for these values of \code{x}.
}
\description{
Transform an lm or glm model that has only two variables into a
function (useful for plotting, see examples).
}
\examples{
data("trees", package = "datasets")
trees_lm1 <- lm(Volume ~ Girth, data = trees)
trees_lm2 <- lm(Volume ~ Girth + I(Girth^2), data = trees)

# Compare these two models on a chart
library(chart)
chart(trees, Volume ~ Girth) +
  geom_point() +
  stat_function(fun = as.function(trees_lm1), col = "red") +
  stat_function(fun = as.function(trees_lm2), col = "blue")

# The created function can also be used for easy predictions
trees_fn1 <- as.function(trees_lm1)
trees_fn1(10:20) # Volume for Girth 10:20
}
