% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.lm.R
\name{chart.lm}
\alias{chart.lm}
\alias{autoplot.lm}
\title{Chart an lm or glm model or diagnose its residuals visually}
\usage{
\method{chart}{lm}(
  data,
  type = "model",
  ...,
  origdata = NULL,
  title,
  labels = "AUTO",
  name = deparse(substitute(data)),
  lang = getOption("SciViews_lang", "en"),
  env = parent.frame()
)

autoplot.lm(
  object,
  origdata = NULL,
  type = c("model", "resfitted", "qqplot", "scalelocation", "cooksd", "resleverage",
    "cookleverage", "reshist", "resautocor"),
  title,
  xlab,
  ylab,
  ...,
  name = deparse(substitute(object)),
  lang = getOption("SciViews_lang", "en"),
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{lm} or \strong{glm} model.}

\item{type}{The type of plot: \code{"model"}, \code{"resfitted"}, \code{"qqplot"},
\code{"scalelocation"}, \code{"cooksd"}, \code{"resleverage"}, \code{"cookleverage"},
\code{"reshist"} or \code{"resautocor"}. For \code{chart()}, can also be provided as
\code{chart$type(....)}.  \code{chart()} also uses \code{"residuals"} that constructs a
combined figure with resfitted, qqplot, scalelocation and
resleverage.}

\item{...}{Additional arguments passed to the chart.``}

\item{origdata}{The original dataset this model was fitted to. Only required
for \code{type = model} and in case untransformed X variable is not in the model.}

\item{title}{A title for the plot. If not provided, a default title is
computed.}

\item{labels}{A vector of four character strings, one for each plot done with
\code{chart$residuals()}.}

\item{name}{The name of the model. If not provided, it is the name of the
model object by default.}

\item{lang}{The language to use for titles and labels, currently only \code{"en"}
or \code{"fr"}.`}

\item{env}{The environment to evaluate code. It is \code{parent.frame()} by
default, and there is no reasons to change it, unless you really know what
you are doing!}

\item{object}{Idem}

\item{xlab}{A label for the X axis. A default label is proposed if it is not
provided.}

\item{ylab}{A label for the Y axis (with default if not provided).}
}
\value{
The ggplot object produced.
}
\description{
The methods \code{autoplot()} or \code{chart()} for \strong{lm} or \strong{glm}
objects. If \code{type = model} (by default for \code{chart()}), a scatterplot with the
model superimposed is produced, providing the model has only two numeric
variables (or a combination of these). The other types allow to analyze the
residuals of the model.
}
\examples{
library(chart)
data(trees, package = "datasets")
trees_lm <- lm(Volume ~ Girth, data = trees)
chart(trees_lm) # origdata not needed because untransformed variables
# Residuals analysis
chart$resfitted(trees_lm)
chart$qqplot(trees_lm)
chart$scalelocation(trees_lm)
chart$cooksd(trees_lm)
chart$resleverage(trees_lm)
chart$cookleverage(trees_lm)
chart$reshist(trees_lm, bins = 15)
chart$resautocor(trees_lm)

# The four most important residual analysis plots in one figure
chart$residuals(trees_lm)

trees_lm2 <-  lm(Volume ~ log(Girth), data = trees)
chart(trees_lm2, origdata = trees) # origdata needed, cf. transformed Girth
trees_lm3 <- lm(Volume ~ Girth + Height, data = trees)
# chart(trees_lm3) # Error because more than 2 variables!
# Polynomial regressions work too
trees_lm4 <- lm(Volume ~ I(Girth^2) + Girth, data = trees)
chart(trees_lm4)
# or using poly()
trees_lm5 <- lm(Volume ~ poly(Girth, 3), data = trees)
chart(trees_lm5, origdata = trees) # origdata required here!
}
