% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.nls.R
\name{chart.nls}
\alias{chart.nls}
\alias{autoplot.nls}
\title{Chart an nls model or diagnose its residuals visually}
\usage{
\method{chart}{nls}(
  data,
  type = "model",
  ...,
  title,
  labels = "AUTO",
  name = deparse(substitute(data)),
  lang = getOption("SciViews_lang", "en"),
  env = parent.frame()
)

autoplot.nls(
  object,
  type = c("model", "resfitted", "qqplot", "scalelocation", "reshist", "resautocor"),
  title,
  xlab,
  ylab,
  ...,
  name = deparse(substitute(object)),
  lang = getOption("SciViews_lang", "en"),
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{nls} model.}

\item{type}{The type of plot: \code{"model"}, \code{"resfitted"}, \code{"qqplot"},
\code{"scalelocation"}, \code{"reshist"} or \code{"resautocor"}. For \code{chart()}, can also
be provided as \code{chart$type(....)}. \code{chart()} also uses \code{"residuals"} that
constructs a combined figure with resfitted, qqplot, scalelocation and
resautocor.}

\item{...}{Additional arguments passed to the chart.}

\item{title}{A title for the plot. If not provided, a default title is
computed.}

\item{labels}{A vector of four character strings, one for each plot done with
\code{chart$residuals()}.}

\item{name}{The name of the model. If not provided, it is the name of the
model object by default.}

\item{lang}{The language to use for titles and labels, currently only \code{"en"}
or \code{"fr"}.}

\item{env}{The environment to evaluate code. It is \code{parent.frame()} by
default, and there is no reasons to change it, unless you really know what
you are doing!}

\item{object}{Idem}

\item{xlab}{A label for the X axis. A default label is proposed if it is not
provided.}

\item{ylab}{A label for the Y axis (with default if not provided).}
}
\value{
The ggplot object produced.
}
\description{
The methods \code{autoplot()} or \code{chart()} for \strong{nls} objects. If
\code{type = model} (by default for \code{chart()}), a scatterplot with the model
superimposed is produced,. The other types allow to analyze the residuals of
the model.
}
\examples{
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]

# Adjust a logistic curve
chick1_logis <- nls(weight ~ SSlogis(Time, Asym, xmid, scal), data = chick1)
library(chart)
chart(chick1_logis)
# Residuals analysis
chart$resfitted(chick1_logis)
chart$qqplot(chick1_logis)
chart$scalelocation(chick1_logis)
chart$reshist(chick1_logis, bins = 15)
chart$resautocor(chick1_logis)

# The four most important residual analysis plots in one figure
chart$residuals(chick1_logis)
}
