% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelit-package.R
\docType{package}
\name{modelit-package}
\alias{modelit}
\alias{modelit-package}
\title{modelit: 'SciViews' - Statistical Models Within 'SciViews::R'}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Create and Use Statistical Models (Linear, General, Nonlinear, ...) by using a SciViews::R common formula interface and label-aware functions.
}
\details{
The {modelit} package provides an extension to base R functions for model
fitting like \code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}} or \code{\link[=nls]{nls()}} with enhanced plots and utilitarian
functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/modelit}
  \item \url{https://www.sciviews.org/modelit/}
  \item Report bugs at \url{https://github.com/SciViews/modelit/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
