% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelr-reexport.R
\name{modelr-reexport}
\alias{modelr-reexport}
\alias{add_predictions}
\alias{add_residuals}
\alias{geom_ref_line}
\alias{rmse}
\alias{mae}
\alias{qae}
\alias{rsquare}
\title{Reexport of modelr functions}
\usage{
add_predictions(data, model, var = "pred", type = NULL)

add_residuals(data, model, var = "resid")

geom_ref_line(h, v, color = "red", colour = color, size = 1)

rmse(model, data)

mae(model, data)

qae(model, data, probs = c(0.05, 0.25, 0.5, 0.75, 0.95))

rsquare(model, data)
}
\arguments{
\item{data}{A data frame}

\item{model}{A model that has a \code{predict()} method.}

\item{var}{A string with the name of the predictions or residuals variable
(by default, it is \code{"pred"} and \code{"resid"} respectively)}

\item{type}{If the model's \code{predict()} method has a \verb{type=} argument, you can
give it here.}

\item{h}{Position of the horizontal reference line}

\item{v}{Position of the vertical reference line}

\item{color}{The color of the reference line}

\item{colour}{Same as above (use the one you prefer)}

\item{size}{The width of the reference line}

\item{probs}{A numeric vector of probabilities}
}
\value{
A function with argument \code{x} that returns the values predicted by the
model for these values of \code{x}.
}
\description{
\code{\link[=add_predictions]{add_predictions()}} and \code{\link[=add_residuals]{add_residuals()}} are pipe-friendly
functions to add predictions or residuals to a data frame.
\code{\link[=geom_ref_line]{geom_ref_line()}} adds a vertical of horizontal reference line. \code{\link[=rmse]{rmse()}}
(the root-mean-squared-error), [mae[]] (the mean absolute error), \code{\link[=qae]{qae()}}
(the quantiles of absolute error) and \code{\link[=rsquare]{rsquare()}} (the variance of the
predictions divided by the variance of the response) are useful model
metrics.
}
\examples{
data(trees, package = "datasets")
trees_lm <- lm(Volume ~ Girth + I(Girth^2), data = trees)
rmse(trees_lm, trees)
rsquare(trees_lm, trees)
mae(trees_lm, trees)
qae(trees_lm, trees, probs = c(0, 0.25, 0.5, 0.75, 1)) # Resids five numbers

add_predictions(trees, trees_lm)
add_residuals(trees, trees_lm)

library(chart)
chart(trees_lm) +
  geom_ref_line(h = 0) # Not particularly useful here, just an example
}
