% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{nls_equation}
\alias{nls_equation}
\title{Extract equation of nls models}
\usage{
nls_equation(x, var_names = NULL, op_latex = c("\\\\cdot", "\\\\times"))
}
\arguments{
\item{x}{The nls object or summary.nls}

\item{var_names}{A named character vector as c("old_var_name" = "new name")}

\item{op_latex}{The operator character to use in fancy scientific notation}
}
\value{
a character value with latex equation
}
\description{
This function try to extract the equation of the Self-starting Nonlinear Models available in {stats}.
}
\examples{
Chick.1 <- ChickWeight[ChickWeight$Chick == 1, ]
fm1 <- nls(weight ~ SSlogis(Time, Asym, xmid, scal), data = Chick.1)
summary(fm1)

nls_equation(fm1)
nls_equation(summary(fm1))

fm1 <- nls(weight ~ SSlogis(Time, Asym = A, xmid = x, scal = scale), data = Chick.1)
summary(fm1)

nls_equation(fm1)
nls_equation(summary(fm1))

nls_equation(summary(fm1), var_names = c("weight" = "Body weight [gm]", "Time" = "Number of days"))

}
