% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_.R
\name{summary_}
\alias{summary_}
\title{Object summaries with the original object as an attribute}
\usage{
summary_(object, ...)
}
\arguments{
\item{object}{An object to be summarized.}

\item{...}{Additional arguments passed to the \code{summary()} function.}
}
\value{
A summary object with an additional \code{"object"} attribute containing
the original input.
}
\description{
This function attempts to summarize an object using the standard \code{\link[base:summary]{base::summary()}}
function. The original object is attached as an attribute to the result for
reference.
}
\examples{

is_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
summary(is_lm)

summary_(is_lm)
attr(summary_(is_lm), "object")

}
