% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{tabularise_coef.summary.nls}
\alias{tabularise_coef.summary.nls}
\title{Tabularise a table of coefficients from summary.nls object}
\usage{
\method{tabularise_coef}{summary.nls}(
  data,
  header = TRUE,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A summary.nls object.}

\item{header}{If TRUE, add a title to the table.}

\item{lang}{The language to use. The default value can be set with, e.g., options(data.io_lang = "fr") for French.}

\item{show.signif.stars}{If TRUE, add the significance stars to the table.}

\item{...}{Additional arguments. Not used.}

\item{env}{The environment where to evaluate formulas (you probably do not need to change the default).}
}
\description{
This method extracts and formats the table of coefficients from summary.nls object, similar to \code{\link[stats:coef]{stats::coef()}}, but in flextable object.
}
\examples{
# example code
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]
# Adjust a logistic curve
chick1_logis <- nls(weight ~ SSlogis(Time, Asym, xmid, scal), data = chick1)
chick1_logis_sum <- summary(chick1_logis)
library(tabularise)
tabularise$coef(chick1_logis_sum)

}
