% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_default.anova}
\alias{tabularise_default.anova}
\title{Create a rich-formatted table from an anova object}
\usage{
\method{tabularise_default}{anova}(data, ...)
}
\arguments{
\item{data}{An \strong{anova} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_tidy.anova]{tabularise_tidy.anova()}}}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Create a rich-formatted table from an anova object
}
\examples{
is.lm <- anova(lm(data = iris, Petal.Length ~ Species))
library(tabularise)
tabularise(is.lm)
}
