% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_default.aov}
\alias{tabularise_default.aov}
\title{Create a rich-formatted table from an aov object}
\usage{
\method{tabularise_default}{aov}(data, ...)
}
\arguments{
\item{data}{An \strong{aov} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_default.anova]{tabularise_default.anova()}}}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions.
}
\description{
Create a rich-formatted table from an aov object
}
\examples{
iris_aov <- aov(data = iris, Petal.Length ~ Species)
tabularise::tabularise$tidy(iris_aov)
}
