% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_default.lm}
\alias{tabularise_default.lm}
\title{Create a rich-formatted table from an lm object}
\usage{
\method{tabularise_default}{lm}(data, ...)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.lm]{tabularise_coef.lm()}}}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Create a rich-formatted table from an lm object
}
\examples{
is.lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
library(tabularise)
tabularise(is.lm)

}
