% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_default.summary.lm}
\alias{tabularise_default.summary.lm}
\title{Create a rich-formatted table from an summary.lm object}
\usage{
\method{tabularise_default}{summary.lm}(data, ..., env = parent.frame())
}
\arguments{
\item{data}{An \strong{summary.lm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.summary.lm]{tabularise_coef.summary.lm()}}}

\item{env}{The environment where to evaluate the model.}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Create a rich-formatted table from an summary.lm object
}
\examples{
is.lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
is.sum <- summary(is.lm)
library(tabularise)
tabularise(is.sum)
}
