% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.glm.R
\name{tabularise_glance.glm}
\alias{tabularise_glance.glm}
\title{Create a glance version of the glm object as a rich-formatted table}
\usage{
\method{tabularise_glance}{glm}(
  data,
  header = TRUE,
  title = header,
  equation = header,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  lang = getOption("data.io_lang", "en"),
  ...,
  kind = "ft",
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{glm} object}

\item{header}{Logical. If \code{TRUE} (\code{TRUE} by default), a header is added to the table.
The header includes both the title and the equation (if applicable).
If set to \code{FALSE}, neither the title nor the equation will be displayed in
the table header, even if the \code{title} or \code{equation} parameters are provided.}

\item{title}{If \code{TRUE} (\code{TRUE} by default), add a title to the table header. Default to the same
value than header, except outside of a chunk where it is \code{FALSE} if a table
caption is detected (\code{tbl-cap} YAML entry).}

\item{equation}{Logical or character. Controls whether an equation is added
to the table header and how parameters are used. Accepted values are:
\itemize{
\item \code{TRUE} (default) : The equation is generated and added to the table
header. Its parameters are also used in the "Term" column.
\item \code{FALSE}: No equation is generated or displayed, and its parameters are
not used in the "Term" column.
\item Character string: A custom equation is provided directly and added to
the table header.
}}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically
from data or \verb{origdata=}.}

\item{origdata}{The original data set this model was fitted to. By default it
is \code{NULL} and no label is used.}

\item{labs}{Labels to change the names of elements in the \code{term} column of
the table. By default it is \code{NULL} and nothing is changed.}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{...}{Additional arguments passed to \code{\link[tabularise:equation]{tabularise::equation()}}}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate formulas (you probably do not
need to change the default).}
}
\value{
A \strong{flextable} object is produced that you can print in different
formats (HTML, LaTeX, Word, PowerPoint) or rearrange with the \{flextable\}
functions.
}
\description{
Turn the glance of \strong{glm} object into a rich-formatted table with
\{flextable\}. The table can be printed in different formats (HTML, LaTeX,
Word, PowerPoint), or rearranged later on.
}
\examples{
iris_glm <- glm(data = iris, Petal.Length ~ Sepal.Length)
tabularise::tabularise$glance(iris_glm)
tabularise::tabularise$glance(iris_glm, equation = FALSE)
tabularise::tabularise$glance(iris_glm, equation = "my personal equation")

}
