% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_glance.lm}
\alias{tabularise_glance.lm}
\title{Tidy version of the lm object into a flextable object}
\usage{
\method{tabularise_glance}{lm}(
  data,
  header = TRUE,
  title = TRUE,
  equation = TRUE,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  lang = getOption("data.io_lang", "en"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{header}{If \code{TRUE} (by default), add a header to the table}

\item{title}{If \code{TRUE} (by default), add a title to the table header}

\item{equation}{If \code{TRUE} (by default), add a equation to the table header. The equation can also be passed in the form of a character string.}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically of the object.}

\item{origdata}{The original dataset this model was fitted to. By default is \code{NULL}}

\item{labs}{Labs to change the names of elements in the \code{term} column of the table. By default is \code{NULL}}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{...}{Additional arguments passed to \code{\link[equatiomatic:extract_eq]{equatiomatic::extract_eq()}}}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Tidy version of the lm object into a flextable object
}
\examples{
is.lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
library(tabularise)
tabularise$glance(is.lm)
}
