% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_tidy.anova}
\alias{tabularise_tidy.anova}
\title{Tidy version of the anova object into a flextable object}
\usage{
\method{tabularise_tidy}{anova}(data, ...)
}
\arguments{
\item{data}{An \strong{anova} object}

\item{...}{Additional arguments used \code{\link[=tabularise_default.anova]{tabularise_default.anova()}}}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions.
}
\description{
Tidy version of the anova object into a flextable object
}
\examples{
is <- data.io::read("iris", package = "datasets")

is_lm1 <- lm(data = is, petal_length ~ species)

library(tabularise)

anova(is_lm1) |> tabularise_tidy()
# identical
anova(is_lm1) |> tabularise$tidy()
# Use labels
anova(is_lm1) |> tabularise$tidy(origdata = is)

# alternative with anova_() in {modelit} package
anova_(is_lm1) |> tabularise$tidy()
}
