% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_tidy.lm}
\alias{tabularise_tidy.lm}
\title{Tidy version of the lm object into a flextable object}
\usage{
\method{tabularise_tidy}{lm}(
  data,
  header = TRUE,
  title = header,
  equation = header,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  conf.int = FALSE,
  conf.level = 0.95,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  kind = "ft"
)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{header}{Logical. If \code{TRUE} (\code{TRUE}by default), a header is added to
the table. The header includes both the title and the equation (if
applicable). If set to \code{FALSE}, neither the title nor the equation will be
displayed in the table header, even if the \code{title} or \code{equation} parameters
are provided.}

\item{title}{If \code{TRUE} (by default) , add a title to the table header.
Default to the same value than header, except outside of a chunk where it is
\code{FALSE} if a table caption is detected (\code{tbl-cap} YAML entry).}

\item{equation}{Logical or character. Controls whether an equation is added
to the table header and how parameters are used. Accepted values are:
\itemize{
\item \code{TRUE}(by default): The equation is generated and added to the table
header. Its parameters are also used in the "Term" column.
\item \code{FALSE}: No equation is generated or displayed, and its
parameters are not used in the "Term" column.
\item \code{NA}: The equation is generated but not displayed in the table header.
Its parameters are used in the "Term" column.
\item Character string: A custom equation is provided directly and added to
the table header.
}}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically
from data or \verb{origdata=}.}

\item{origdata}{The original data set this model was fitted to. By default it
is \code{NULL} and no label is used.}

\item{labs}{Labels to change the names of elements in the \code{term} column of
the table. By default it is \code{NULL} and no term is changed.}

\item{conf.int}{If \code{TRUE}, add the confidence interval. The default is
\code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int = TRUE}. The default is 0.95.}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE}, add the significance stars to the table.
The default is \code{getOption("show.signif.stars")}}

\item{...}{Additional arguments passed to \code{\link[tabularise:equation]{tabularise::equation()}}}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}
}
\value{
A \strong{flextable} object that you can print in different formats (HTML,
LaTeX, Word, PowerPoint) or rearrange with the \{flextable\} functions.
}
\description{
Create a rich-formatted table with the 'tidy' information from an \strong{lm}
object.
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
tabularise::tabularise$tidy(iris_lm)
}
