% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_.R
\name{lm_}
\alias{lm_}
\title{Fitting Linear Models by SciViews}
\usage{
lm_(data, formula, ...)
}
\arguments{
\item{data}{A data frame}

\item{formula}{An object of class formula}

\item{...}{All the arguments of the lm() function}
}
\value{
an lm_ object if attribute additions have been made. Otherwise, the object will be of class lm
}
\description{
The lm_() function is used like the lm() function from the {stats} package. It allows adding additional elements such as labels and units.
}
\examples{
data(iris)
res <- lm_(iris, formula = Petal.Length ~ Sepal.Length + Species)
res
class(res)

}
