% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelit-package.R
\docType{package}
\name{modelit-package}
\alias{modelit-package}
\title{Statistical Models for 'SciViews::R'}
\description{
The \{modelit\} package provides an extension to base R functions for model
fitting like \code{\link[=lm]{lm()}}, \code{\link[=glm]{glm()}} or \code{\link[=nls]{nls()}} with enhanced plots and utilitarian
functions.
}
\section{Important functions}{

\itemize{
\item \code{\link[=fit_model]{fit_model()}} creates a \strong{model_fit} object that has many methods.
\item \code{\link[=tabularise]{tabularise()}} methods for \strong{lm}, \strong{glm}, \strong{nls}, \strong{model_fit},
\strong{anova} and \strong{aov} objects.
\item \code{\link[=chart]{chart()}} methods for \strong{lm}, \strong{glm}, \strong{nls} and \strong{model_fit} objects.
\item \code{\link[=as.function]{as.function()}} transforms an \strong{lm} or \strong{nls} model into a function that
can be plotted using \code{stat_function()}.
}
}

