% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelit-package.R
\docType{package}
\name{modelit-package}
\alias{modelit}
\alias{modelit-package}
\title{modelit: Statistical Models within SciViews::R}
\description{
Create and Use Statistical Models (Linear, General, Nonlinear, ...)
  by using a SciViews::R common formula interface and label-aware functions.
}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
