% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_.R
\name{summary.lm_}
\alias{summary.lm_}
\title{Summarizing Linear Model Fits by SciViews}
\usage{
\method{summary}{lm_}(object, ...)
}
\arguments{
\item{object}{an object of class lm_}

\item{...}{additional argument to stats:::summary.lm()}
}
\value{
an object of class summary.lm_ object, similar to summary.lm
}
\description{
summary method for class lm_
}
\examples{
#TODO
}
