% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{tabularise_coef.nls}
\alias{tabularise_coef.nls}
\title{Create a rich-formatted table using the coefficients of the nls object}
\usage{
\method{tabularise_coef}{nls}(
  data,
  header = TRUE,
  equation = header,
  lang = getOption("data.io_lang", "en"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{nls} object.}

\item{header}{If \code{TRUE} (by default), add a title to the table.}

\item{equation}{If \code{TRUE} (by default), add the equation of the model}

\item{lang}{The language to use. The default value can be set with, e.g., \code{options(data.io_lang = "fr")} for French.}

\item{...}{Additional arguments.}

\item{env}{The environment where to evaluate lazyeval expressions (unused for now).}
}
\value{
A \strong{flextable} object you can print in different forms or rearrange
with the {flextable} functions.
}
\description{
This method extracts and formats the coefficients from nls object, similar to \code{\link[stats:coef]{stats::coef()}}, but in flextable object.
}
\examples{
# example code
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]
# Adjust a logistic curve
chick1_logis <- nls(weight ~ SSlogis(Time, Asym, xmid, scal), data = chick1)
library(tabularise)
tabularise$coef(chick1_logis)
}
