% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.glm.R
\name{tabularise_coef.summary.glm}
\alias{tabularise_coef.summary.glm}
\title{Create a rich-formatted table using the table of coefficients of the summary.glm object}
\usage{
\method{tabularise_coef}{summary.glm}(data, ..., env = parent.frame())
}
\arguments{
\item{data}{A \strong{summary.glm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_tidy.glm]{tabularise_tidy.glm()}}}

\item{env}{The environment where to evaluate the model.}
}
\value{
A \strong{flextable} object that you can print in different formats
(HTML, LaTeX, Word, PowerPoint) or rearrange with the {flextable}
functions.
}
\description{
Create a rich-formatted {flextable} object with the table of coefficients
from the \code{\link[=summary]{summary()}} of a \strong{glm} object.
}
\examples{
iris_glm <- glm(data = iris, Petal.Length ~ Sepal.Length)
iris_glm_sum <- summary(iris_glm)
tabularise::tabularise$coef(iris_glm_sum)
}
