% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_coef.summary.lm}
\alias{tabularise_coef.summary.lm}
\title{Create a rich-formatted table using the table of coefficients of the summary.lm object}
\usage{
\method{tabularise_coef}{summary.lm}(data, ..., kind = "ft", env = parent.frame())
}
\arguments{
\item{data}{A \strong{summary.lm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_tidy.lm]{tabularise_tidy.lm()}}}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate the model.}
}
\value{
A \strong{flextable} object you can print in different formats (HTML,
LaTeX, Word, PowerPoint) or rearrange with the {flextable} functions.
}
\description{
Create a rich-formatted table using the table of coefficients of the summary.lm object
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
iris_lm_sum <- summary(iris_lm)
tabularise::tabularise$coef(iris_lm_sum)
}
