% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{tabularise_coef.summary.nls}
\alias{tabularise_coef.summary.nls}
\title{Create a rich-formatted table using the table of coefficients of the summary.nls object}
\usage{
\method{tabularise_coef}{summary.nls}(
  data,
  header = TRUE,
  title = header,
  equation = header,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{summary.nls} object.}

\item{header}{If \code{TRUE} (by default), add a title to the table.}

\item{title}{If \code{TRUE} (by default), add a title to the table header}

\item{equation}{Add equation of the model to the table. If \code{TRUE},
\code{\link[=equation]{equation()}} is used. The equation can also be passed in the form of a
character string (LaTeX equation).}

\item{lang}{The language to use. The default value can be set with, e.g.,
\code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE} (by default), add the significance stars
to the table.}

\item{...}{Additional arguments passed to \code{\link[=equation]{equation()}}}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object that you can print in different forms or
rearrange with the {flextable} functions.
}
\description{
This function extracts and formats the table of coefficients from a
\strong{summary.nls} object, similar to \code{\link[stats:coef]{stats::coef()}}, but in flextable object.
}
\examples{
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]
# Adjust a logistic curve
chick1_logis <- nls(data = chick1, weight ~ SSlogis(Time, Asym, xmid, scal))
chick1_logis_sum <- summary(chick1_logis)

tabularise::tabularise$coef(chick1_logis_sum)
tabularise::tabularise$coef(chick1_logis_sum, header = FALSE, equation = TRUE)
}
