% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.glm.R
\name{tabularise_default.glm}
\alias{tabularise_default.glm}
\title{Create a rich-formatted table from an glm object}
\usage{
\method{tabularise_default}{glm}(
  data,
  lang = getOption("data.io_lang", "en"),
  footer = TRUE,
  ...
)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{lang}{The natural language to use. The default value can be set with, e.g., \code{options(data.io_lang = "fr")} for French.}

\item{footer}{If \code{TRUE} (by default), add a footer to the table}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.glm]{tabularise_coef.glm()}}}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Create a rich-formatted table from an glm object
}
\examples{
is.lm <- glm(data = iris, Petal.Length ~ Sepal.Length)
library(tabularise)
tabularise(is.lm)
}
