% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_default.lm}
\alias{tabularise_default.lm}
\title{Create a rich-formatted table from an lm object}
\usage{
\method{tabularise_default}{lm}(data, ..., kind = "ft", env = parent.frame())
}
\arguments{
\item{data}{An \strong{lm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.lm]{tabularise_coef.lm()}}}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate the model.}
}
\value{
A \strong{flextable} object that you can print in different formats (HTML,
LaTeX, Word, PowerPoint) or rearrange with the {flextable} functions.
}
\description{
The default \code{\link[=tabularise]{tabularise()}} method for \strong{lm} objects create a minimalist
table with result of the analysis in a rich-formatted tabular presentation.
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
tabularise::tabularise(iris_lm)
}
