% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.glm.R
\name{tabularise_default.summary.glm}
\alias{tabularise_default.summary.glm}
\title{Create a rich-formatted table from a summary.glm object}
\usage{
\method{tabularise_default}{summary.glm}(data, ..., env = parent.frame())
}
\arguments{
\item{data}{A \strong{summary.glm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.summary.glm]{tabularise_coef.summary.glm()}}}

\item{env}{The environment where to evaluate the model.}
}
\value{
A \strong{flextable} object that you can print in different form or
rearrange with the {flextable} functions.
}
\description{
Create a rich-formatted table version of the \code{\link[=summary]{summary()}} of a \strong{glm} object.
}
\examples{
iris_glm <- glm(data = iris, Petal.Length ~ Sepal.Length)
iris_glm_sum <- summary(iris_glm)
tabularise::tabularise(iris_glm_sum)
}
