% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.glm.R
\name{tabularise_default.summary.glm}
\alias{tabularise_default.summary.glm}
\title{Create a rich-formatted table from an summary.glm object}
\usage{
\method{tabularise_default}{summary.glm}(data, ..., env = parent.frame())
}
\arguments{
\item{data}{An \strong{summary.glm} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.summary.glm]{tabularise_coef.summary.glm()}}}

\item{env}{The environment where to evaluate the model.}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Create a rich-formatted table from an summary.glm object
}
\examples{
is.lm <- glm(data = iris, Petal.Length ~ Sepal.Length)
is.sum <- summary(is.lm)
library(tabularise)
tabularise(is.sum)
}
