% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_default.summary.lm}
\alias{tabularise_default.summary.lm}
\title{Create a rich-formatted table from an summary.lm object}
\usage{
\method{tabularise_default}{summary.lm}(
  data,
  footer = TRUE,
  lang = getOption("data.io_lang", "en"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{summary.lm} object}

\item{footer}{If \code{TRUE} (by default), add a footer to the table}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{...}{Additional arguments passed to \code{\link[=tabularise_coef.summary.lm]{tabularise_coef.summary.lm()}}}

\item{env}{The environment where to evaluate the model.}
}
\value{
A \strong{flextable} object you can print in different formats (HTML,
LaTeX, Word, PowerPoint) or rearrange with the {flextable} functions.
}
\description{
Create a rich-formatted table from an summary.lm object
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
iris_lm_sum <- summary(iris_lm)
tabularise::tabularise(iris_lm_sum)
}
