% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{tabularise_default.summary.nls}
\alias{tabularise_default.summary.nls}
\title{Tabularise a summary.nls object into a flextable object}
\usage{
\method{tabularise_default}{summary.nls}(
  data,
  header = TRUE,
  footer = TRUE,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A summary.nls object}

\item{header}{If TRUE add a title to the table}

\item{footer}{If TRUE add a footer to the table}

\item{lang}{The language to use. The default value can be set with, e.g., options(data.io_lang = "fr") for French.}

\item{show.signif.stars}{If TRUE add the significance stars to the table}

\item{...}{Unused argument}

\item{env}{The environment where to evaluate the function}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
This method tabularises a summary.nls object as a flextable.
}
\examples{
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]

# Adjust a logistic curve
chick1_logis <- nls(weight ~ SSlogis(Time, Asym, xmid, scal), data = chick1)
library(tabularise)
tabularise(summary(chick1_logis))

}
