% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.nls.R
\name{tabularise_default.summary.nls}
\alias{tabularise_default.summary.nls}
\title{Create a rich-formatted table from a summary.nls object}
\usage{
\method{tabularise_default}{summary.nls}(
  data,
  header = TRUE,
  footer = TRUE,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{A \strong{summary.nls} object.}

\item{header}{If \code{TRUE} (by default), add a header to the table}

\item{footer}{If \code{TRUE} (by default), add a footer to the table.}

\item{lang}{The language to use. The default value can be set with, e.g.
\code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE} (by default), add the significance stars
to the table.}

\item{...}{Additional arguments (Not used).}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object that you can print in different forms or
rearrange with the {flextable} functions.
}
\description{
Create a table of a \strong{summary.nls} object. This table looks like the output
of \code{\link[=print.summary.nls]{print.summary.nls()}} but richly formatted. The \code{\link[=tabularise_coef]{tabularise_coef()}}
function offers more customization options for this object.
}
\examples{
data("ChickWeight", package = "datasets")
chick1 <- ChickWeight[ChickWeight$Chick == 1, ]

# Adjust a logistic curve
chick1_logis <- nls(data = chick1, weight ~ SSlogis(Time, Asym, xmid, scal))
chick1_logis_sum <- summary(chick1_logis)

tabularise::tabularise(chick1_logis_sum)
tabularise::tabularise(chick1_logis_sum, footer = FALSE)
}
\seealso{
\code{\link[tabularise:tabularise]{tabularise::tabularise()}}, \code{\link[tabularise:tabularise_tidy]{tabularise::tabularise_tidy()}},
\code{\link[=tabularise_coef.summary.nls]{tabularise_coef.summary.nls()}}
}
