% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_glance.lm}
\alias{tabularise_glance.lm}
\title{Glance version of the lm object into a flextable object}
\usage{
\method{tabularise_glance}{lm}(
  data,
  header = TRUE,
  title = NULL,
  equation = TRUE,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  lang = getOption("data.io_lang", "en"),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{header}{If \code{TRUE} (by default), add a header to the table}

\item{title}{If \code{TRUE}, add a title to the table header. Default to the same
value than header, except outside of a chunk where it is \code{FALSE} if a table
caption is detected (\code{tbl-cap} YAML entry).}

\item{equation}{If \code{TRUE} (by default), add a equation to the table header.
The equation can also be passed in the form of a character string (LaTeX).}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically
from data or \verb{origdata=}.}

\item{origdata}{The original data set this model was fitted to. By default it
is \code{NULL} and no label is used (only the name of the variables).}

\item{labs}{Labels to change the names of elements in the \code{term} column of
the table. By default it is \code{NULL} and nothing is changed.}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{...}{Additional arguments passed to \code{\link[tabularise:equation]{tabularise::equation()}}}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object that you can print in different form or
rearrange with the {flextable} functions.
}
\description{
Create a rich-formatted table with the 'glance' information from an \strong{lm}
object.
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
tabularise::tabularise$glance(iris_lm)
}
