% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_tidy.anova}
\alias{tabularise_tidy.anova}
\title{Tidy version of the anova object into a flextable object}
\usage{
\method{tabularise_tidy}{anova}(
  data,
  header = TRUE,
  title = header,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{anova} object}

\item{header}{If \code{TRUE} (by default), add a header to the table}

\item{title}{If \code{TRUE} (by default), add a title to the table header}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically of the object}

\item{origdata}{The original dataset this model was fitted to. By default is \code{NULL}}

\item{labs}{Labs to change the names of elements in the \code{term} column of the table. By default is \code{NULL}}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE} (by default), add the significance stars to the table.}

\item{...}{Additional arguments (unused for now)}

\item{env}{The environment where to evaluate lazyeval expressions (unused for now)}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Tidy version of the anova object into a flextable object
}
\examples{
is.lm <- anova(lm(data = iris, Petal.Length ~ Species))
library(tabularise)
tabularise$tidy(is.lm)
}
