% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_tidy.anova}
\alias{tabularise_tidy.anova}
\title{Tidy version of the anova object into a flextable object}
\usage{
\method{tabularise_tidy}{anova}(data, ...)
}
\arguments{
\item{data}{An \strong{anova} object}

\item{...}{Additional arguments used tabularise_default.anova()}
}
\value{
A \strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions.
}
\description{
Tidy version of the anova object into a flextable object
}
\examples{
iris_anova <- anova(lm(data = iris, Petal.Length ~ Species))
tabularise::tabularise$tidy(iris_anova)
}
