% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_tidy.aov}
\alias{tabularise_tidy.aov}
\title{Tidy version of the aov object into a flextable object}
\usage{
\method{tabularise_tidy}{aov}(data, ..., kind = "ft", env = parent.frame())
}
\arguments{
\item{data}{An \strong{anova} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_tidy.anova]{tabularise_tidy.anova()}}}

\item{kind}{The kind of table to produce: "tt" for tinytable, or "ft" for
flextable (default).}

\item{env}{The environment where to evaluate the object.}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the {flextable} functions.
}
\description{
Tidy version of the aov object into a flextable object
}
\examples{
iris_aov <- aov(data = iris, Petal.Length ~ Species)
tabularise::tabularise$tidy(iris_aov)
}
