% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.anova.R
\name{tabularise_tidy.aov}
\alias{tabularise_tidy.aov}
\title{Tidy version of the aov object into a flextable object}
\usage{
\method{tabularise_tidy}{aov}(data, ...)
}
\arguments{
\item{data}{An \strong{aov} object}

\item{...}{Additional arguments passed to \code{\link[=tabularise_default.anova]{tabularise_default.anova()}}}
}
\value{
\strong{flextable} object you can print in different form or rearrange
with the \{flextable\} functions.
}
\description{
Tidy version of the aov object into a flextable object
}
\examples{
iris_aov <- aov(data = iris, Petal.Length ~ Species)
tabularise::tabularise$tidy(iris_aov)
}
