% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabularise.lm.R
\name{tabularise_tidy.lm}
\alias{tabularise_tidy.lm}
\title{Tidy version of the lm object into a flextable object}
\usage{
\method{tabularise_tidy}{lm}(
  data,
  header = TRUE,
  title = header,
  equation = header,
  auto.labs = TRUE,
  origdata = NULL,
  labs = NULL,
  conf.int = FALSE,
  conf.level = 0.95,
  lang = getOption("data.io_lang", "en"),
  show.signif.stars = getOption("show.signif.stars", TRUE),
  ...,
  env = parent.frame()
)
}
\arguments{
\item{data}{An \strong{lm} object}

\item{header}{If \code{TRUE} (by default), add an header to the table}

\item{title}{If \code{TRUE} (by default), add a title to the table header}

\item{equation}{If \code{TRUE} (by default), add an equation to the table header.
The equation can also be passed in the form of a character string (LaTeX).}

\item{auto.labs}{If \code{TRUE} (by default), use labels (and units) automatically
from \verb{origdata=}.}

\item{origdata}{The original data set this model was fitted to. By default it
is \code{NULL} and no label is used.}

\item{labs}{Labels to change the names of elements in the \code{term} column of
the table. By default it is \code{NULL} and no term is changed.}

\item{conf.int}{If \code{TRUE}, add the confidence interval. The default is
\code{FALSE}.}

\item{conf.level}{The confidence level to use for the confidence interval if
\code{conf.int = TRUE}. The default is 0.95.}

\item{lang}{The natural language to use. The default value can be set with,
e.g., \code{options(data.io_lang = "fr")} for French.}

\item{show.signif.stars}{If \code{TRUE}, add the significance stars to the table.
The default is \code{getOption("show.signif.stars")}}

\item{...}{Additional arguments passed to \code{\link[equatiomatic:extract_eq]{equatiomatic::extract_eq()}}}

\item{env}{The environment where to evaluate lazyeval expressions (unused for
now).}
}
\value{
A \strong{flextable} object that you can print in different formats (HTML,
LaTeX, Word, PowerPoint) or rearrange with the {flextable} functions.
}
\description{
Create a rich-formatted table with the 'tidy' information from an \strong{lm}
object.
}
\examples{
iris_lm <- lm(data = iris, Petal.Length ~ Sepal.Length)
tabularise::tabularise$tidy(iris_lm)
}
